+++ packages/dashboard/app/scripts/views/transcriptView.js
@@ -136,7 +136,9 @@ haDash.Views = haDash.Views || {};
         })
       .fail(function(jqXHR, textStatus, errorThrown) {
         console.log(errorThrown);
-        alert('Error: ' + textStatus + '\n' + errorThrown);
+        alert("We have trouble aligning your media, your file may be too large or in a format we don't understand. Sorry.");
+        this.$el.find('button.align').show();
+        this.$el.find('button.aligning').hide();
       });
       // spin
       //$('#mediaDetail').data('view').refresh();
@@ -145,6 +147,8 @@ haDash.Views = haDash.Views || {};
 
     delete: function() {
       if (this.notMutable()) return;
+
+      if (!confirm('Are you sure you wish to delete this transcript?')) return;
 
       this.model.destroy({
         url: haDash.API + '/transcripts/' + this.model.id,
